requirements.txt
Flask==2.0.3
keras==2.8.0rc1
numpy==1.21.4
opencv_python==4.5.4.58
Pillow==9.0.1
tf_bodypix==0.3.9
